fix: XML parse boolean
