+++ .changeset/gorgeous-ligers-give.md
@@ -1,0 +1,5 @@
+---
+'@bnb-chain/greenfield-js-sdk': patch
+---
+
+fix: XML boolean parse
