+++ packages/chain-sdk/src/clients/spclient/spApis/getUserBuckets.ts
@@ -1,4 +1,5 @@
 import { GetUserBucketsResponse } from '@/types';
+import { convertStrToBool } from '@/types/sp-xml/Common';
 import xml from 'xml2js';
 
 // https://docs.bnbchain.org/greenfield-docs/docs/api/storgae-provider-rest/get_user_buckets
@@ -18,7 +19,8 @@ export const parseGetUserBucketsResponse = async (data: string) => {
     Buckets = Buckets.map((item) => {
       return {
         ...item,
-        Removed: Boolean(item.Removed),
+        // @ts-ignore
+        Removed: convertStrToBool(item.Removed),
         DeleteAt: Number(item.DeleteAt),
         UpdateAt: Number(item.UpdateAt),
         UpdateTime: Number(item.UpdateTime),
